requirements.txt
flask
requests
beautifulsoup4
pillow
gunicorn
